ATS_resume_analyzer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
openpyxl==3.1.2
aiosmtplib==3.0.1
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
